requirements.txt
diff-match-patch==20121119
Django==2.0.7
django-crispy-forms==1.7.2
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.0.1
django-pure-pagination==0.3.0
django-reversion==3.0.0
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
jdcal==1.4
jieba==0.39
Markdown==2.6.11
mysqlclient==1.3.13
odfpy==1.3.6
openpyxl==2.5.4
Pygments==2.2.0
pytz==2018.5
PyYAML==3.13
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
